+++ CustomCraftingStations/ModEntry.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using ChroniclerCherry.Common.Integrations.GenericModConfigMenu;
 using CustomCraftingStations.Framework;
 using Microsoft.Xna.Framework;
 using StardewModdingAPI;
@@ -45,6 +46,7 @@ internal class ModEntry : Mod
             return;
         }
 
+        I18n.Init(helper.Translation);
         this.Config = helper.ReadConfig<ModConfig>();
 
         helper.Events.GameLoop.GameLaunched += this.OnGameLaunched;
@@ -110,6 +112,12 @@ internal class ModEntry : Mod
     private void OnGameLaunched(object sender, GameLaunchedEventArgs e)
     {
         this.CookingSkillMenu = Type.GetType("CookingSkill.NewCraftingPage, CookingSkill");
+
+        this.AddGenericModConfigMenu(
+            new GenericModConfigMenuIntegrationForCustomCraftingStations(),
+            get: () => this.Config,
+            set: config => this.Config = config
+        );
     }
 
     /// <inheritdoc cref="IInputEvents.ButtonPressed" />
@@ -151,7 +159,12 @@ internal class ModEntry : Mod
 
         Vector2 centeringOnScreen = Utility.getTopLeftPositionForCenteringOnScreen(800 + IClickableMenu.borderWidth * 2, 600 + IClickableMenu.borderWidth * 2);
 
-        Game1.activeClickableMenu = new CustomCraftingMenu((int)centeringOnScreen.X, (int)centeringOnScreen.Y, 800 + IClickableMenu.borderWidth * 2, 600 + IClickableMenu.borderWidth * 2, chests, station.CraftingRecipes, station.CookingRecipes);
+        bool isCooking = station.CookingRecipes.Count > 0;
+        var recipes = isCooking
+            ? station.CookingRecipes
+            : station.CraftingRecipes;
+
+        Game1.activeClickableMenu = new CustomCraftingMenu((int)centeringOnScreen.X, (int)centeringOnScreen.Y, 800 + IClickableMenu.borderWidth * 2, 600 + IClickableMenu.borderWidth * 2, chests, recipes, isCooking);
     }
 
     private void OpenAndFixMenu(IClickableMenu instance)