+++ CustomCraftingStations/docs/release-notes.md
@@ -1,6 +1,13 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## 1.4.0
+Released 25 January 2025 for SMAPI 4.1.10 or later. Updated by Pathoschild.
+
+- Added in-game config UI through [Generic Mod Config Menu](https://www.nexusmods.com/stardewvalley/mods/5098).
+- Added validation error if a custom station has both cooking and crafting recipes, which isn't supported by the vanilla menu used since 1.2.0.
+- Fixed cooking recipes in recent versions.
+
 ## 1.3.0
 Released 21 January 2025 for SMAPI 4.1.10 or later. Updated by Pathoschild.