+++ CustomizeAnywhere/docs/release-notes.md
@@ -1,6 +1,12 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## 1.4.0
+Released 25 January 2025 for SMAPI 4.1.10 or later. Updated by Pathoschild.
+
+- Added in-game config UI through [Generic Mod Config Menu](https://www.nexusmods.com/stardewvalley/mods/5098).
+- Improved translations. Thanks to BlackRosePetals (added Chinese)!
+
 ## 1.3.0
 Released 21 January 2025 for SMAPI 4.1.10 or later. Updated by Pathoschild.