+++ UpgradeEmptyCabins/docs/release-notes.md
@@ -1,4 +1,9 @@
 ﻿# Release notes
+## 1.2.1
+Released 25 January 2025 for SMAPI 4.1.10 or later. Updated by Pathoschild.
+
+* Improved translations. Thanks to BlackRosePetals (added Chinese)!
+
 ## 1.2.0
 Released 21 January 2025 for SMAPI 4.1.10 or later. Updated by Pathoschild.